+++ fiddles/jquery/fiddle-0041-FabricImageCloud/src/meta.js
@@ -1,0 +1,5 @@
+app.metadata = app.metadata || {
+    consoleTag: 'H O U S E ~ f i d d l e s',
+    gitHubUrl: 'https://github.com/bradyhouse/house/tree/master/fiddles/jquery/fiddle-0041-fabricImageCloud',
+    dataUrl: 'data.json'
+  };