+++ fiddles/jquery/fiddle-0041-FabricImageCloud/src/toolkit/canvas.js
@@ -1,6 +1,6 @@
 /**
  * Wrapper class for the fabric.Canvas.
- * @link {http://fabricjs.com/docs/fabric.Canvas.html}
+ * @see http://fabricjs.com/docs/fabric.Canvas.html
  */
 
 class Canvas extends Base {
@@ -8,7 +8,11 @@ class Canvas extends Base {
   config() {
     return {
       hook: 'canvas1',
-      children: null
+      children: [],
+      backgroundColor: 'rgb(100,100,200)',
+      width: 500,
+      height: 500,
+      onImageClick: null
     };
   }
 
@@ -17,45 +21,62 @@ class Canvas extends Base {
     if (config) {
       this.apply(this, config, this.config());
     }
+    this._start = 0;
+    this._stop = 0;
     this.init();
   }
 
-  /**
-   * Method used to append the docElement to
-   * configured hook element.
-   */
-  bind() {
-    if (this.hook && this.docElementNS) {
-      this.hook.appendChild(this.docElementNS);
-    }
+  get start() {
+    return this._start;
   }
 
-  /**
-   * Method called by the constructor to create
-   * and assign docElement based
-   * on the properties exposed by the class.
-   *
-   * Note - if the autoBind flag is true,
-   * then it ends by invoking bind method.
-   */
+  set start(value) {
+    this._start = value;
+  }
+
+  get stop() {
+    return this._stop;
+  }
+
+  set stop(value) {
+    this._stop = value;
+  }
+
   init() {
+    var me = this,
+      canvas = new fabric.Canvas(this.hook, {
+        width: this.width,
+        height: this.height
+      });
 
+    if (canvas) {
+      if (this.children && this.children.length) {
+        this.children.map(function (child) {
+          if (child.fabric) {
+            // ToDo: Report Bug
+            child.fabric.setHeight(child.height);
+            canvas.add(child.fabric);
+          }
+        }, canvas);
+      }
+      canvas.on('mouse:down', function(object) {
+        var date = new Date();
+        me.start = date.getTime();
+      });
 
-
-
-    if (this.children && this.children.length) {
-      this.children.map((child) => {
-        child.parent = this;
-        if (child.docElementNS) {
-          this.docElementNS.appendChild(child.docElementNS);
+      canvas.on('mouse:up', function (object) {
+        var date = new Date(),
+            image = object && object.target && object.target.type && object.target.type === 'image' ? object.target : null,
+            delta = 0;
+        me.stop = date.getTime();
+        delta = me.stop - me.start;
+        if (delta < 200 && image) {
+          if (me.hasOwnMethod('onImageClick')) {
+            me['onImageClick'].call(me, image);
+          }
         }
       });
-    }
-
-    if (this.autoBind) {
-      this.bind();
+      this.fabric = canvas;
     }
   }
-
-
 }