+++ fiddles/nativeScript/fiddle-0000-Template/nativescript/app/shared/core/index.ts
@@ -1,7 +1,0 @@
-// services
-export * from './services/modal-native.service';
-export * from './services/window-native.service';
-export * from './services/http.service';
-
-// utils
-export * from './utils/actionbar.util';